Resolve requires before checking
Summary:
Before this diff, we would link requires by first creating unresolved tvars for
every require, then resolving those tvars to a ModuleT type. We would create a
tvar for each import/require, which can be more than the number of actual
dependencies, in cases where a dependency is required more than once.

The tvar juggling is an artifact of the old, global inference system in Flow,
where we didn't always have a resolved type for every dependency. Since types-
first, we always have a resolved type for dependencies, even in the presence of
cycles.

So, this diff simplifies things by collecting the dependency module's types
before we start checking a file. We pass the resolved dependency types to the
checked file's context.

This eliminates the "add_require_tvar" and "connect requires" steps from type
checking. We replace it with much simpler code to define a "resolve require"
function which we pass to `Context.make`.

With this change, we no longer need file sig for type checking. We only used it
to get the locations of each require. (We still need the file sig for other
things, like type normalization, coverage, and references.)

Reviewed By: SamChou19815

Differential Revision: D47608370

fbshipit-source-id: a4c8585f9c7418b237f66216e7c38b2dea456b9f