[flow][module-resolution] Make .js.flow only shadow .js file if they are in the same directory
Summary:
To prepare for multiplatform support with .js.flow file as the common interface file, we want to cleanup the current weird behavior of .js.flow first. Under haste, it can shadow a corresponding .js file from anywhere in the codebase, which can be very surprising.

This diff will start to error when we find that a .js file if we find that it's shadowed by a .js.flow file without the same path prefix.

Changelog: [error] Under Haste module system, it will now be an error if a `.js.flow` file shadows a `.js` file with a different path prefix.

Reviewed By: samwgoldman

Differential Revision:
D47455131

------------------------------------------------------------------------
(from c97f1f2d4be36f5a10212a6409d43862d8bd1d8b)

fbshipit-source-id: 09e48557d2747c05d43deab06c985d879207e394