[flow][rename] Actually support global rename
Summary:
Currently, the global rename did all the hard work of globally finding references, but only rename the references in the file that initiates the request...

This diff actually performs renames on all files that contain the found references, and added test to test the behavior.

The test currently shows identical set of results as find-refs modulo json format differences. The test exposes something we don't do right currently. e.g. in the case of

```
// a.js
export const a = 3;
// b.js
import {a as b} from './a'
```

rename of `a` will produce

```
// a.js
export const NEW_NAME = 3;
// b.js
import {a as NEW_NAME} from './a'
```

which is wrong, but it's not the job of this diff to fix it. I will followup in a future diff.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D47446644

fbshipit-source-id: 451e4b0f5fce50c41e0054e3c13f8906253a4978