[flow][rename] More conservative global rename behavior compared to find all references
Summary:
As established in the previous diff, global rename needs to be more conservative than find all references. This diff implements the idea.

During find all refs when we look for all references of starting from imports, we should only keep the reference of imports, and do not find more local references of the import. This requires us to update the rename mapper so that

```
import {a} from '...'
```

becomes

```
import {NEW_NAME as a} from '...'
```

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D47604217

fbshipit-source-id: 5c81103bb19840c5bb769d20fd2f39905f953659