[flow][refactor][rename] Add `FindRefsTypes.request` type and pass it around
Summary:
Global renames should be more conservative than global find references. e.g.

```
// a.js
export const a = 3;
// b.js
import {a} from './a';
export {a};
```

Global find refs might highlight all `a`, but global rename should only change it to:

```
export const NEW_NAME = 3;
// b.js
import {NEW_NAME as a} from './a';
export {a};
```

To achieve this, we need to customize the find-all-ref results based on the kind. This diff setup the types and propagate the request around, but doesn't do any customization yet.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D47604219

fbshipit-source-id: ed254dd66fa6a6ddde9933b3ffa06bd3385b828c