[flow][ide] Include find ref results from jsx props
Summary:
This diff fixes a bug where we miss the find all ref results from jsx props. In a codebase when most of the JS code is related to react, that will be a big shortcoming.

We achieve this by visiting the jsx props during candidate search, and perform a component_type ~> props object conversion. The result object type's id should be already included in the obj_to_obj map, so the result can be found.

Changelog: [ide] Fixed a bug where local find references do not return references in jsx props

Reviewed By: samwgoldman

Differential Revision: D47736407

fbshipit-source-id: 4b5bcfac5c5945c0d97ffda2e592a725aa809496