[flow] Model what a component renders in AbstractComponent
Summary:
This diffs adds a third type argument to AbstractComponent to model what a component returns. This allows us to describe more properties about components in types, and in turn, enforce those properties in code.

This third type argument is covariantly subtyped, drawing from the intuition that the `renders` method return type and function component return types are both in covariant positions.

I separated out test cases specifically for the new behavior into the next diff to make it easier to review the .exp files. The changes to tests in this diff are to add the third type argument when `React$AbstractComponent` is used directly.

Reviewed By: mvitousek

Differential Revision: D47305225

fbshipit-source-id: 54004226ee1cf0c16300dd04f96ec9b3040fc8cb