[flow] Make the second type argument to React$AbstractComponent truly optional
Summary:
This stack adds a third type argument to AbstractComponent to model what a component renders and introduces "Render Types" for component syntax.

React$AbstractComponent currently requires two type arguments. React.AbstractComponent (exported in the React libdef) provides a default for the second type argument. We're adding a third type argument to model what a component renders, but there are enough direct usages of React$AbstractComponent that this is difficult to roll out. To ease the rollout pain, I'm adding default type arguments to React$AbstractComponent directly.

In order to continue to give good error messages for a synthetic type created with no good source location, I added a new reason that refers to the nth type argument of an arbitrary type.

Reviewed By: mvitousek

Differential Revision: D47760369

fbshipit-source-id: ecfb07a875cf1aa143a6d7026b9c4fc759c9104f