[flow][get-def] Do not fail the request for unresolvable remote def_loc for imports
Summary:
When we encounter unresolvable remote def_loc for imports, like

```
import {foo} from 'haha_i_do_not_exist'
```

it's not a logical error of Flow. In this case, we should just return the current location as the def_loc. Currently, returning an error makes no visual differences for get-def requests, but it will make a rename request fail in the future when the rename starts to depend on this code.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D47813873

fbshipit-source-id: 0576ec3d32dd820a47d9f5e53fec1b2980a519e0