[flow] Switch to using ReactAbstractComponentT for components
Summary: This diff makes the switch from modeling components as FunTs to modeling them as ReactAbstractComponentTs. As part of doing so, we need to extend implicit instantiation--and a couple other parts of the system--to handle PolyT(ReactAbstractComponentT) types, which would not have occurred previously. Fortunately, these turn out to be pretty straightforward. Hopefully this diff handles every such case--I audited all appearances of PolyT in flow's implementation, but I could have misread a case that's relevant and thought it was irrelevant.

Reviewed By: jbrown215

Differential Revision: D46872716

fbshipit-source-id: b286ae8e8aa496f50809c62c70ff642f62d6ef4e