[flow][docs] Update options page
Summary:
Make some updates to the options page. I will not attempt to document all currently non-documented options, of which there are many. But at least we can not list no longer valid options in the index. Also add `enums` option to the list.

Options are now shown in the right sidebar, without having to manually update a TOC. Deprecated options are moved to their own section.

Additionally, the type of the option is moved out of the title to clean that up.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D48053821

fbshipit-source-id: 7497f8dda498cba5aa3320d32c49bf046902dfac