[flow] categorize get-def on jsx
Summary:
get-def on non-identifier jsx element names and attribute names isn't currently supported. right now it just doesn't find the location in the AST. we want to properly classify all requests, so that we can be sure we're returning empty responses in legitimate places.

this one is not a legitimate place, but fixing that is a TODO.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D48116965

fbshipit-source-id: 14235d6aa47450147725bedb567ffafd18bcf247