[flow] categorize get-def on type params
Summary:
get-def on type params isn't currently supported. right now it just doesn't find the location in the AST. we want to properly classify all requests, so that we can be sure we're returning empty responses in legitimate places.

it should point at itself (triggering find-refs, to show the uses), but fixing that is a TODO.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D48117779

fbshipit-source-id: ca506ed6710138d21e02ecfaee3835696e715445