[flow] categorize get-def on object keys
Summary:
get-def on non-identifier object keys isn't currently supported. right now it just doesn't find the location in the AST. we want to properly classify all requests, so that we can be sure we're returning empty responses in legitimate places.

this one is not a legitimate place, but fixing that is a TODO.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D48116966

fbshipit-source-id: bfb35dd4c7310d7b7ad9c38aab37766a2dbb69cf