[flow][try] Automatically derive Try Flow option docs links from name
Summary:
We can automatically derive the Try Flow options docs link from the name of the option.
Additionally, the Try Flow config in `flow_dot_js.ml` is per Flow version, better to compute the actual link url in the website itself so if we change the url format we can change how we compute it.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D48132021

fbshipit-source-id: 047754529bc9c9d4e1456c6bd99b4a2f39aea444