[flow][env_builder] Model `exports` as a regular global
Summary:
Currently, `exports` has a weird setup. We model it as a special global write, and in type_inference_js, immediately resolve it to `{-[string]: mixed}`. I believe the weird setup is a leftover of unsealed objects, so I cleaned it up in this diff by making it a regular global.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D48277308

fbshipit-source-id: 69a6e7d6931fd4b1b9eaf185b60d5df9aa566732