[flow][test][EZ] Rename test
Summary:
The requirement isn't that js.flow and the file has to be in the same directory, but also that they have to have the same base name once the .flow is stripped. In the next diff, it will slightly changed so that `a.js.flow` shadowing `a.android.js` will be allowed. Therefore, I just rename the test suite as `js_flow_illegal_shadow`

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D48363595

fbshipit-source-id: 40c77670dfde4b8395a1da616a52002df61bf02a