[flow][multi-plat] First cut of multiplatform support
Summary:
In this diff, I added an initial support for multiplatform support, currently designed to check a react-native codebase with .ios.js and .android.js files under one flowconfig. For such pattern, we will require that a common interface file in .js.flow exist. Both implementation must subtype against the common interface, and the common interface will ensure that both implementations exist.

This is achieved through the concept of synthetic imports. For implementation files, we synthesize a self import and an import of the interface file, and for the interface file, we will synthesize an import of implementation files. In this diff, I used the types of the synthesized import to verify that the module type of implementation files are subtypes of the module type of the common interface file. The check that both implementations exist have not been implemented yet.

I added tests for both Haste and Node module system to demonstrate that the subtyping part of the system works, and Flow can generate reasonably good error messages. There are still a lot of to, but don't have to be done in this diff, since everything is well gated:

- Ensure that when we have an common interface file, we can only have (.js file) xor (.ios.js && .android.js) file

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D47387567

fbshipit-source-id: 70d45d6d2956fac52892eecc018ada57373c3059