[flow][multi-plat] Pick a location that feels like signature location to be the module conformance error loc with best effort
Summary:
Currently when there is a module conformance error, we will error on the entire file, which will turn the entire file red in IDEs. This is terrible UX.

However, there is no obvious locations to error on, but we should still do our best. I come up with the following rules that would probably handle 99% of the cases:

- For cjs, we will first try to pick the location of module.exports, then fallback to the first module.exports prop assignment
- For esm, we will first try to pick the location of default exports, then fallback to the first export.

If a file doesn't have any export, then we will fallback to the entire file, but that probably won't happen a lot.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision:
D48374923

------------------------------------------------------------------------
(from 23d919a167184cfcd116569605e6d4e3679a60cf)

fbshipit-source-id: f199f0bfcd0f3816dbe58b94e65fb7862bb98138