[flow][autocomplete] Improve string literal types in indexed access autocomplete
Summary:
This diff builds on top of the function factored out in D48527716.

However, it's not sufficient to call that function alone. We have a function `fix_loc_of_string_token`, which will change the location of the string literal token to only the first character. This is added to detect the cases like

```
Foo['      + ''
```

where the parser might consider the string token to be `'      + '`.

This diff adds a special case that when the token is `''` or `""`, we return the original token location, since it's much more likely that `''` is a locally complete token. With that, the workflow becomes

```
Obj['']
     ^ autocomplete
edit_locs = (insert loc between quotes, replace loc of literal)
==> fix_loc_of_string_token doesn't change edit locs
==> autocomplete_create_string_literal_edit_controls changes the edit locs to be (loc of literal, loc of literal)
==> replace '' with 'foo', profit
```

Changelog: [ide] We fixed a bug that causes us to insert extra quote in autocomplete string literal in indexed access types

Reviewed By: samwgoldman

Differential Revision: D48527714

fbshipit-source-id: ca8c6b881d642506abd2568dcc64f6222f83766b