[flow] Default `module.system.haste.use_name_reducers` to true
Summary:
This is the starting point of deleting support for `providesModule`, by first no longer enabling it by default if you use Haste. This will allow us to remove the options from all the usages first.

Changelog: [breaking] `module.system.haste.use_name_reducers` now defaults to true. We intend to remove the option to configure it to false in the future. You will only be impacted if you use `module.system=haste`, which is unlikely outside of Meta.

Reviewed By: samwgoldman

Differential Revision: D48486482

fbshipit-source-id: f1bdb08fcc0fe121ec8fa1aae6f890b28abe24a0