[flow][tests] Re-record a few haste package related tests with name_reducers on
Summary:
It doesn't help a lot to test the providesModule behavior, since it's rarely used. This diff switches it to use name reducer, so that providesModule can be killed in the future.

I have to add some haste excludes to preserve the original behavior that some files are not under haste, so most tests don't require re-recording. The suite `haste_node_package_and_file` is deleted, since the edge case it tests no longer applies.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D48568828

fbshipit-source-id: 57383e22870bfe95dbd5e575907b027dd5bc7509