[flow] Delete abnormal handling code for statements
Summary:
Cleans up abnormal handling code for statements. `statement.ml` is now much cleaner! Delete the separate `Toplevels` module.

Abnormals are not fully deleted, as they are still used in the typing of expressions, and `Statement.expression` continues to throw (though the only abnormal now left is `Throw` as it is the only one that can occur in expressions).

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D48591754

fbshipit-source-id: 73d06a700be236042627efc19ca46062f0af03a7