[flow][render-types] Make component ID optional
Summary:
Per discussion in Flow Core, we're expanding render types to be able to represent nodes that are not components.

The `RendersT {id = None, super}` representation is used for the `$Renders<super>` syntax (better syntax will come). It lets the type checker know that we're checking against a render type and not simply super.

This diff just changes the structure-- the changes are used later in the stack.

Reviewed By: mvitousek

Differential Revision: D48371404

fbshipit-source-id: 7c13643ad4db37df63ccc1e04a4173262395d4a4