[flow] Turn elements of components into their render types
Summary:
This diff adds a ReactPromoteRendersRepresentation destructor that:
1. Given an object whose type field is a named component, turn into the render type for that component
2. Given a named component, turn into the render type for that component

This function is what lets us turn `$Renders<React.Element<Foo>>` into the render type for `Foo`.

In order to give developers good ergonomics, we've chosen to also allow `$Renders<Foo>` to mean the same thing. This means that $Renders works over:
1. Named components
2. Any subtype of React.Node

There is no case in which a named component would have an ambiguous render type because components are not subtypes of React.Node.

It's still not clear if we need to check that the argument is a subtype of React.Node except for in component returns. I will add those checks as necessary after more investigation.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D48400801

fbshipit-source-id: c9ff2b556fdedc6beab19499f626b6b2be64257f