[flow] Components are component types in type positions
Summary:
We've decided that `Foo` is no longer the render type for component `Foo` and is instead the type of Foo when used as a type.

This allows for `React.Element<Foo>`.

As a fun side-effect, we also require (and allow) you to monomorphize components when used directly as a type, which will help solve several issues with polymorphic components + type destructors

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D48398713

fbshipit-source-id: 79e5b25b3dd0ca046441a84cf08875a8e871dd5a