[flow][tests] Change `declaration_files_haste` and `declaration_files_incremental_haste` test suite to use haste name reducers
Summary:
It doesn't help a lot to test the providesModule behavior, since it's rarely used. This diff switches it to use name reducer, so that providesModule can be killed in the future.

I deleted all the ExplicitProvidesModule files and their imports, since haste with name reducers would no longer use them. The other expected change is that `SomeOtherModule` no longer resolves. Previously it can resolve due to the providesModule pragma in the source file, not due to package.json. The package.json in node_modules/qux/corge2 is not a valid one, since it's not directly under `node_modules`.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D48666993

fbshipit-source-id: 5b091fc6000fb69daf4ce7f45f4c4f859a2ca125