[flow][tests] Change `node_haste` test suite to use haste name reducers
Summary:
It doesn't help a lot to test the providesModule behavior, since it's rarely used. This diff switches it to use name reducer, so that providesModule can be killed in the future.

I have to add some haste excludes to preserve the original behavior that some files are not under haste, so most tests don't require re-recording.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D48666994

fbshipit-source-id: fbe02c01c58f6eeb224e4cd819c8d23e798835b3