[flow][tests] Change `parse_error_haste` test suite to use haste name reducers
Summary:
It doesn't help a lot to test the providesModule behavior, since it's rarely used. This diff switches it to use name reducer, so that providesModule can be killed in the future.

I changed the import so that client no longer imports the providesModule name. Also updated the comments, since it's no longer accurate even before this test change.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D48666995

fbshipit-source-id: 146b13dae095e71aa8f0ec601e3348cb3c42be6d