[flow] Nominal component normalization under render types
Summary:
When we encounter a nominal AbstractComponent under a render type we should print it as its component name. This ensures that `renders Foo` is printed as `renders Foo` and not `renders AbstractComponent<...>`.

I gate this to only under a render type because it is otherwise more informative to see the config and other fields.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D48664926

fbshipit-source-id: 7428d1b8c308d04b32a1368cb72abce6e6111125