[flow][tests] Change `ensure_parsed_rename` to use haste name reducers
Summary:
It doesn't help a lot to test the providesModule behavior, since it's rarely used. This diff switches it to use name reducer, so that providesModule can be killed in the future.

We keep testing the same behavior by placing the files with the same names in different directories.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D48666992

fbshipit-source-id: 04e6559642b32cc5b47032a18838d121de4ad35b