[flow][tests] Change `incremental_delete` to use haste name reducers
Summary:
It doesn't help a lot to test the providesModule behavior, since it's rarely used. This diff switches it to use name reducer, so that providesModule can be killed in the future.

The notable change to make sure it still tests the same thing is that we moved dupe1.js and dupe2.js that both provide Dupe via pragma into separate directories and rename both of them as Dupe.js

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D48666997

fbshipit-source-id: fbc4bc33eb742d7306066ada5f8588d51607f44a