[flow][tests] Change `incremental_cycle_break` to use haste name reducers
Summary:
It doesn't help a lot to test the providesModule behavior, since it's rarely used. This diff switches it to use name reducer, so that providesModule can be killed in the future.

This test suite starts with a cyclic requires of two haste modules A and B, and then replaces `A` with a node module `A` by removing provideModules, and then restore it.
The same incremental behavior can be exercised by deleting and restoring A in the haste name reducer world.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D48666996

fbshipit-source-id: 0ed486ed969394ee8a82f7ed3619f9d3ee78f935