[flow][tests] Delete `incremental_mixed_naming_cycle` test suite
Summary:
In this test, we add a comment to a module that's in a strongly connected component of cyclic modules. It's known that comments changes are no-op under types first, so this test probably doesn't have a lot of value. This diff deletes it, and it's the last test that uses `providesModule`

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D48706242

fbshipit-source-id: 7b0f98bba94c412c3b4d76b963cd9f696467b88e