[flow] Kill `module.system.haste.use_name_reducers=false`
Summary:
With all internal repo and test usages of this option gone, and unlikely external usage, we can kill the option for good.

Changelog: [breaking] The ability to configure `module.system.haste.use_name_reducers=false` is removed. For now this option defaults to true and can only be set to true. The option to set the only possible value will be removed in a future release.

Reviewed By: samwgoldman

Differential Revision: D48666986

fbshipit-source-id: d969f48e92c6a5135bf1fe04464c27c581956754