[flow][render-types] Polymorphic Element Renders Promotion
Summary:
ObjT promotion when the component is polymorphic is difficult-- the Element does not keep track of the type arguments that were used to instantiate the component.

We can't add an additional field that is not actually present at runtime to model this feature. Instead, we use the prop map's id to associate the object with the monomorphized component and use that prop map to look it up when promoting.

We take not to record the monomorphized result by adding a flag to createElement. This also allows us to choose *not* to waste space in the context by keeping track of components that were never polymorphic.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D48728111

fbshipit-source-id: e68ddbb5e6d3f2c109fda293143ce2d2b0a6f7c6