[flow][render-types] Add context field to keep track of monomorphized components from jsx
Summary:
ObjT promotion when the component is polymorphic is difficult-- the Element does not keep track of the type arguments that were used to instantiate the component.

We can't add an additional field that is not actually present at runtime to model this feature. Instead, we use the prop map's id to associate the object with the monomorphized component and use that prop map to look it up when promoting.

This diff only implements the Context changes. The next diff implements support for polymorphic ObjT promotion.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D48713844

fbshipit-source-id: 9697fcc1309308942dc0425d3c9b34d987f463aa