[flow][multiplat] Fallback to exported types when we failed to find other exports for module sig location
Summary:
Since there is no obvious location to error on for conformance errors, we pick one with best effort from exports. We didn't include export types by mistake, and we should include it as a fallback, so that we don't error on the entire module when there are some type exports but no value exports.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D48781757

fbshipit-source-id: 8c84fe23b20839121115e02153f7a1ce5854d0ce