[flow][website] Add link for blog posts that are just Medium links
Summary:
It is possible (e.g. through intern search) to go to the page of a blog post that is really hosted on Medium:
https://flow.org/blog/2021/09/13/TypeScript-Enums-vs-Flow-Enums/

However, as you see above, there is no link in that page to the Medium post.

Fix this by just changing the check to see if there is a Medium link.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D48850405

fbshipit-source-id: 6b52c742e1732dd27c4b0e9218fef6af6c65343c