[flow] Thread tparams_map to component types
Summary: Previously we didnt thread the contextual tparams into mk_component in type_annotation. This was probably because it's always empty for declared components, but that's not the case for component types. This might not surface any behavioral changes currently, but it means that substitutions might not work out correctly (because in the absence of the tparams map, we'll read tparams from the environment, getting a GenericT wrapped in a tvar rather than a bare GenericT)

Reviewed By: SamChou19815

Differential Revision: D48907985

fbshipit-source-id: 497ec9f7329536d74305b7f07dfa6e7e630edfff