[flow] Flow abstract component configs
Summary:
This diff addresses an issue with the interaction of AbstractComponentT and generics, which was surfaced by component syntax but didn't originate with it. Previously we flowed mixed when we extract the props from an AbstractComponentT using ElementProps<>: this was because the props type may legitimately be empty, and flowing that type would allow arbitrary things to be read from element props. However, this meant that we would lose contextual typing needed to instantiate generic children based on the children type of their parents -- see the added example. This diff fixes by flowing the props type if its *not* empty, and mixed if it is.

This has the result of creating new incompatibilities, because while previously `React.Element<typeof SomeAbstractComponent>` would have `mixed` as its config, and would accept arbitrary props as lower bounds, now we produce the correct config type, and require incoming elements to have configs that match.

Changelog: [new errors] React.ElementProps is now more precise when applied to AbstractComponents, resulting in the config type of the component rather than `mixed`. This will cause errors when elements of components with different props are passed into `React.Element<>` types.

Reviewed By: SamChou19815

Differential Revision: D48910857

fbshipit-source-id: 7902286f86c75544e4ab132486605451adb96cd9