[flow][multiplat] Error location improvement
Summary:
In this diff, the error location of conformance errors in the multiplatform world is greatly improved. To helper give reviewers more context, I will introduced the error location system and the multiplatform error location first.

## How do we decide an error location

Consider the following program:

```
const o: {+foo: {+bar: [+baz: string]}} = {foo: {bar: [1]}};
```

Ideally, we would like the checker to error on `1`, which is the smallest possible error location that makes sense. How do we achieve that?

We do that by keeping track of these locations in frames of use_ops. In the above example, the final use_op might look something like:

```
TupleElementCompat(lower=<loc of `1`>, upper=<loc of `baz` in typedef>) {
  PropCompat(lower=<loc of `bar: [1]` in object literal>, upper=<loc of `bar: [+baz: string]}` in typedef>) {
    PropCompat(lower=<loc of the entire object literal>, upper=<loc of entire typedef>) {
      // root
      AssignOp(lower=<loc of the entire object literal>, upper=<loc of entire typedef>)
    }
  }
}
```

During error printing, we will find the smallest location via `Loc.contains` on each level, with the assumption that the inner wrapped use_op will generally fully enclose an outer one.

## The locations for modules in multiplatform world

In the multiplatform world, we need to do a module type subtyping between an implementation file and a common interface file. What should be the location of the module type of the implementation file? Logically, it should be the entire module's location.

However, when the optimization mentioned in the above section doesn't work out, we might end up making the entire file red in the IDE. Therefore, we decide to pick a location of an export to represent the signature location of the module, so that we can have some guarantee that we won't turn the entire file red. Unfortunately, this clever trick completely destroys the optimization described in the section above, since the assumption of the enclosing relation of locations are broken.

## The solution

We want the best of the both worlds:
- find the smallest error location where we can
- when we can't, don't make the entire file red.

Fortunately, we can implement this logic in flow_error.ml, we just need to give it enough context:

1. In the root use_op for conformance checks, we record both the location of the entire file, and the smaller signature location of one of the export.
2. We try to find the smallest location using the location of the entire file as an initial input.
3. We check the result of step 2
    - We get a location that's not the location of the entire file. This means that we found a smaller location. Use that location as the error location.
    - We get back the location of the entire file. In this case, use the signature location, so that we don't turn the entire file red.

Changelog: [internal] multiplatform support is not announced externally yet

Reviewed By: samwgoldman

Differential Revision:
D48837080

------------------------------------------------------------------------
(from d8706a82b17392828d47e5c88dd337f9623efb36)

fbshipit-source-id: e647e5a9271301992dc9af0537a6a97122ed1f1f