[flow] Fix exponential blowup in get-def
Summary:
This diff alters the DFS so that:
1. The seen set is mutable
2. We keep track of results for given locations

The reason we need both structures instead of just the result set is because we recurse after updating the seen set but before knowing the solution.

We use the mutable seen set to avoid visiting locations we've already visited in other branches. Before reporting a cyclic error, we check to see if we already have a good solution for this location handy.

This diff also restores the limit check.

Changelog: [bug-fix] Fixed cases where hovering over a parser error could cause the Flow server to appear to hang

Reviewed By: mroch, SamChou19815

Differential Revision:
D48912287

------------------------------------------------------------------------
(from 896be0c48a219e344bd2c0cb8fcce4fa9c24dbe8)

fbshipit-source-id: bf23c1176360b2c66721a53b03b04191b3c2fea3