[flow][tests] Add tests that asserts the current broken documentHighlight and find-ref behavior
Summary:
Currently find-refs are completely broken for variables related to namespace imports. Consider

```
import * as Foo from 'foo';
Foo;
```

Find ref on either Foo will trigger a go-to-def on `Foo` in namespace imports, which jumps to the source file. This final jump is a `Get_def_request.Type`, which currently doesn't have name information, which will trigger the error message which I fixed the grammar in the previous diff.

In addition, the namespace import loc is also not included in the LocalImportRefSearcher, so that the local results in the file will be completely missing.

It will be fixed in the next few diffs.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D48858796

fbshipit-source-id: f9c56d033df506ded830608cfeb9652cefe33512