[flow][multiplat] Support declaring matching classes in common interface files
Summary:
Currently, it's impossible to write a class in the common interface file that can pass conformance check, since it will have a different nominal id as the one declared in implementation file. For react components using classes, we can mostly workaround with `React.AbstractComponent`, but we don't have good workaround in most other cases.

This diff adds support for writing classes in common interface files that can actually pass type check. We will check the file in the pair of class type to see whether we have a pair of matching files (platform-specific impl file, corresponding common interface file), and the same name. If so, we will do a structural check, including:

- class instance shape match
- class statics shape match
- type arguments match
- recurse down the inheritance chain

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D48742282

fbshipit-source-id: 1ec9947995901a4ebfe74958159657ee8ba383b1