[flow][render-types] Less brittle "does not render" matching in error messages
Summary:
I think the current heuristics of deciding whether we turn the "A is incompatible with B" error message into "A doesn't render B" is quite brittle. It relies on reason descriptions, which can be easily overridden, like what I did in the previous diff.

This diff switches it to use use_op to keep track of the information. I argue that we have a pretty good idea when we should do it: when `x ~> RenderT y` happens, it means a renders type check, so we give it a special use_op, and later pattern match on this frame in error printing.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D48995662

fbshipit-source-id: 52a650a8cafff6a3dde0ecba653a8f7a0c6192e7