[flow][find-refs] Fix local find refs for namespace imported variable
Summary:
This diff fixes two issues that make find-refs on namespaced imported variables unusable:

1. We now propagate the name information in Get_def_request.Type request where we can.
2. We include the namespace import location in LocalImportRefSearcher.

With this fix, the local find refs feature for namespaced imported variables is fixed, and the global one no longer errors, but still unfixed. The reason and the fix is more involved and unrelated to this diff, so I will do it the next diff.

Changelog: [fix] We fixed local find references for namespace imported variables. It will no longer error or return empty results.

Reviewed By: samwgoldman

Differential Revision: D48858795

fbshipit-source-id: f796131cf8a403dedea3939a85cc710b47ec9db3