[flow] Propagate deep read only on reads
Summary: This diff begins the process of actually doing something with the react_dro flag. In this diff, whenever we read from an object or array with the dro flag set, we then apply the ReactDRO type destructor to the result, which flips the flag on the result if applicable.

Reviewed By: jbrown215

Differential Revision: D48809516

fbshipit-source-id: 9019a9bcf83d85c89795d8f91438d17acc78b925