[flow] Unbreak refs by allowing .current to be mutable
Summary: The behavior of react refs is that the .current property can be mutated from within useEffect or action handlers. Flow does not currently track whether or not an operation is in such a scope, so by default DRO would ban such behavior. This diff alters it to allow this behavior universally.

Reviewed By: jbrown215

Differential Revision: D48809752

fbshipit-source-id: 65df936bfe54269d12ec87fe7b4b2801f73e14dd