[flow] Error on writes to deep read only
Summary:
Finally, some testable code. Here we raise an error whenever we write to an object or array with the DRO flag set. "writing" here means going through SetProp or SetElem, not e.g. array.push.

We also introduce a utility type to apply the DRO destructor so that we can actually test code!

Reviewed By: jbrown215

Differential Revision: D48809544

fbshipit-source-id: 77591a5f4c1445b401b7a5c6faa449955e7e3192