[flow] Don't wrap instantiation tvars in unions in annots
Summary: Currently, when a union flows into a ReposUseT, we wrap its elements in AnnotTs. This has always been kind of bogus, but in the post-LTI world it's both unnecessary and problematic during implicit instantiation. This diff changes it to only wrap AnnotTs that are not instantiable or when not during implicit instantiation.

Reviewed By: SamChou19815

Differential Revision: D48785396

fbshipit-source-id: 56c1806dd1df4ae1b541635347e524912c1d8357