[flow] Ref extraction
Summary: This diff extracts the instance type from the "ref" prop of syntax components. It does so using a new implicit instantiation tool, where we flow `React$RefSetter<T>` into the type of the ref prop and try to pin `T`. This models calling `forwardRef` on a function component definition as previously typical.

Reviewed By: SamChou19815

Differential Revision: D48922581

fbshipit-source-id: fb7f11d4cd292c58872684c4a255f3f5272b3abd