[flow] Make `React.Element` opaque but still allowing inspection for now
Summary:
This is the diff that makes `React.Element` opaque. It makes error messages a lot nicer, because we are no longer referring to individual props of the `React.Element` object that people really shouldn't be care about, but instead says "type arguments of `React.Element`" doesn't match.

We will keep the record monomorphized component type, but record it via the object type in the upper bound of the opaque type.

Changelog: [errors] `React.Element` is now opaque, but we still expose some of its internal structure via the bound on the opaque type. You should not depend on the internal structure. In a future release, we plan to make `React.Element` truly opaque.

Reviewed By: jbrown215

Differential Revision: D48893055

fbshipit-source-id: 48c9197aba3a607e675527d10fe7758251f9800b